requirements.txt
CSXCAD
numpy
matplotlib
pylab
openEMS-python
mpl_toolkits
meshio
h5py
xml